refactor: update api folder structure
