+++ src/api/rest/exportList/exportList.ts
@@ -1,5 +1,5 @@
 import { commonGet } from '~api/common/commonGet';
-import type { ExportListResponse } from '~api/types/responses/ExportListResponse';
+import type { ExportListResponse } from '~api/types/backend/responses/ExportListResponse';
 
 export const getListFileEndpointById = (listId: string) =>
 	commonGet<ExportListResponse>(`/export-order/generate/${listId}`);
