+++ src/api/rest/lists/dto/transformActiveListsQuantityDTO.ts
@@ -1,6 +1,6 @@
-import type { GetHelpLists } from '~api/types/GetHelp/GetHelpLists';
-import type { HelpCategories } from '~api/types/GetHelp/HelpCategories';
-import type { ListQuantityResponse } from '~api/types/responses/ListQuantityResponse';
+import type { ListQuantityResponse } from '~api/types/backend/responses/ListQuantityResponse';
+import type { GetHelpLists } from '~api/types/getHelp/GetHelpLists';
+import type { HelpCategories } from '~api/types/getHelp/HelpCategories';
 
 export function transformActiveListsQuantityDTO(lists: ListQuantityResponse[]): GetHelpLists {
 	const categories: Record<string, Record<string, HelpCategories>> = {
